+++ lib/src/utils/listenable/listenable_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
 /// {@category utils}
-base class ListenableView<T extends Listenable> implements Listenable {
+class ListenableView<T extends Listenable> implements Listenable {
   final T _listenable;
 
   /// Wraps [listenable] so that only listening is reachable through it.